Convert to use puppeteer
puppeteer has the advantage that it includes a compatible chrome headless by default. This makes the deployment and orchestration far simpler as it means that all dependencies can be included within a single docker image.

Because of this references to docker-compose and docker generally can be removed, especially since it will run locally very happily with only `npm install && npm start`