Make existing test suite resilient to ordering
My depending on stub call order, or out-out-of-scope variable values tests can become dependent on process state that they do not control, and so are sensitive to ordering.

Replace proxyquire with direct stubs

When transitive module dependencies are stubbed with proxyquire there is no easy way to restore them for later tests, so once the module has been stubbed it cannot be unstubbed. This makes the test runner entirely dependent on execution order.

Instead use sinon directly to stub/spy on the relevant modules, which can easily be restored after the appropriate tests have executed.