Add pdf options (#19)
* Prefix debug namespace with `pdf`

This allows all debug for this module to be switched on without needing to set `DEBUG=*` which also turns on debug for other modules - in particular express router and mocha, which make a lot of noise while running unit tests.

* Make existing test suite resilient to ordering

My depending on stub call order, or out-out-of-scope variable values tests can become dependent on process state that they do not control, and so are sensitive to ordering.

Replace proxyquire with direct stubs

When transitive module dependencies are stubbed with proxyquire there is no easy way to restore them for later tests, so once the module has been stubbed it cannot be unstubbed. This makes the test runner entirely dependent on execution order.

Instead use sinon directly to stub/spy on the relevant modules, which can easily be restored after the appropriate tests have executed.

* Add pdf rendering options to API

Allows a client to define what options it needs to pass to chrome's `pdf` method to allow for customised rendering by supporting a `pdfOptions` object on the request body.