Add unhandled rejections flag
Chrome occasionally crashes on teardown and emits an unhandled promise rejection warning. When this occurs it leaves some resources in memory. Eventually this results in running out of memory and failed requests.

By adding this flag the process will restart completely if and when the unhandled promise rejection occurs, and will therefore not continue to consume excess memory.